navigation.ui: move NavigationItemActionListBox to subpackage
