+++ src/main/java/net/kemitix/ldapmanager/ui/NavigationPanel.java
@@ -29,7 +29,7 @@ import com.googlecode.lanterna.gui2.Borders;
 import com.googlecode.lanterna.gui2.LinearLayout;
 import com.googlecode.lanterna.gui2.Panel;
 import lombok.extern.java.Log;
-import net.kemitix.ldapmanager.navigation.NavigationItemActionListBox;
+import net.kemitix.ldapmanager.navigation.ui.NavigationItemActionListBox;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
