+++ src/test/java/net/kemitix/ldapmanager/ui/ChangeOuIT.java
@@ -27,7 +27,7 @@ package net.kemitix.ldapmanager.ui;
 import net.kemitix.ldapmanager.ldap.AbstractLdapConnectionIntegrationTest;
 import net.kemitix.ldapmanager.ldap.LdapNameUtil;
 import net.kemitix.ldapmanager.ldap.LdapOptions;
-import net.kemitix.ldapmanager.navigation.NavigationItemActionListBox;
+import net.kemitix.ldapmanager.navigation.ui.NavigationItemActionListBox;
 import net.kemitix.ldapmanager.state.CurrentContainer;
 import org.junit.Before;
 import org.junit.Test;
